src/demo/requirements.txt
python-binance>=1.0.16
websockets>=10.3
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.5.0
seaborn>=0.12.0
pyyaml>=6.0
python-dotenv>=0.21.0
asyncio>=3.4.3
aiohttp>=3.8.0